Fix filtering by status
