+++ src/app/helpers/listBrowsing.ts
@@ -1,0 +1,16 @@
+export const sortDescending = 'desc';
+export const sortAscending = 'asc';
+export const pageInfo = {
+  pageSize: 10,
+  cursor: '0',
+  page: 1,
+};
+
+export interface SortInfo {
+  column: string;
+  direction: string;
+}
+
+export function getSortingExpression(sort: SortInfo): string {
+  return `${sort.column}:${sort.direction}`;
+}
