--- src/app/modules/orders/orders.constants.ts
@@ -1,7 +1,0 @@
-export const sortDescending = 'desc';
-export const sortAscending = 'asc';
-export const pageInfo = {
-  pageSize: 6,
-  cursor: '0',
-  page: 1,
-};
